+++ src/ParallelAlgorithms/LinearSolver/Gmres/ResidualMonitorActions.hpp
@@ -71,18 +71,16 @@ struct InitializeResidualMagnitude {
         iteration_id, residual_magnitude, residual_magnitude};
 
     // Do some logging
-    if (UNLIKELY(static_cast<int>(
-                     get<logging::Tags::Verbosity<OptionsGroup>>(cache)) >=
-                 static_cast<int>(::Verbosity::Verbose))) {
+    if (UNLIKELY(get<logging::Tags::Verbosity<OptionsGroup>>(cache) >=
+                 ::Verbosity::Verbose)) {
       Parallel::printf("Linear solver '" +
                            Options::name<OptionsGroup>() +
                            "' initialized with residual: %e\n",
                        residual_magnitude);
     }
     if (UNLIKELY(has_converged and
-                 static_cast<int>(
-                     get<logging::Tags::Verbosity<OptionsGroup>>(cache)) >=
-                     static_cast<int>(::Verbosity::Quiet))) {
+                 get<logging::Tags::Verbosity<OptionsGroup>>(cache) >=
+                     ::Verbosity::Quiet)) {
       Parallel::printf("The linear solver '" +
                            Options::name<OptionsGroup>() +
                            "' has converged without any iterations: %s\n",
@@ -190,18 +188,16 @@ struct StoreOrthogonalization {
         get<initial_residual_magnitude_tag>(box)};
 
     // Do some logging
-    if (UNLIKELY(static_cast<int>(
-                     get<logging::Tags::Verbosity<OptionsGroup>>(cache)) >=
-                 static_cast<int>(::Verbosity::Verbose))) {
+    if (UNLIKELY(get<logging::Tags::Verbosity<OptionsGroup>>(cache) >=
+                 ::Verbosity::Verbose)) {
       Parallel::printf("Linear solver '" +
                            Options::name<OptionsGroup>() +
                            "' iteration %zu done. Remaining residual: %e\n",
                        completed_iterations, residual_magnitude);
     }
     if (UNLIKELY(has_converged and
-                 static_cast<int>(
-                     get<logging::Tags::Verbosity<OptionsGroup>>(cache)) >=
-                     static_cast<int>(::Verbosity::Quiet))) {
+                     get<logging::Tags::Verbosity<OptionsGroup>>(cache) >=
+                     ::Verbosity::Quiet)) {
       Parallel::printf("The linear solver '" +
                            Options::name<OptionsGroup>() +
                            "' has converged in %zu iterations: %s\n",
